organised a little
fixed two item
