+++ Assets/Prefabs/Environment/PlayerUnlockable/HeavyUnlock.prefab
@@ -56,6 +56,7 @@ GameObject:
   m_Component:
   - component: {fileID: 6231977346123998344}
   - component: {fileID: 7195286101846185183}
+  - component: {fileID: 4129984187387917410}
   m_Layer: 0
   m_Name: Spot Light
   m_TagString: Untagged
@@ -140,6 +141,29 @@ Light:
   m_UseViewFrustumForShadowCasterCull: 1
   m_ShadowRadius: 0
   m_ShadowAngle: 0
+--- !u!114 &4129984187387917410
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 5504081414000004636}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 474bcb49853aa07438625e644c072ee6, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  m_Version: 3
+  m_UsePipelineSettings: 1
+  m_AdditionalLightsShadowResolutionTier: 2
+  m_LightLayerMask: 1
+  m_RenderingLayers: 1
+  m_CustomShadowLayers: 0
+  m_ShadowLayerMask: 1
+  m_ShadowRenderingLayers: 1
+  m_LightCookieSize: {x: 1, y: 1}
+  m_LightCookieOffset: {x: 0, y: 0}
+  m_SoftShadowQuality: 0
 --- !u!1 &5523386548212966826
 GameObject:
   m_ObjectHideFlags: 0
@@ -150,6 +174,7 @@ GameObject:
   m_Component:
   - component: {fileID: 5320785212622099613}
   - component: {fileID: 7926073440948716257}
+  - component: {fileID: 9191465689202455366}
   m_Layer: 0
   m_Name: HeavyUnlock
   m_TagString: UnlockableHammer
@@ -195,6 +220,19 @@ SphereCollider:
   serializedVersion: 3
   m_Radius: 0.75
   m_Center: {x: 0, y: 0, z: 0}
+--- !u!114 &9191465689202455366
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 5523386548212966826}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 747f8daff34ce5249af7ce003539deb8, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  abilityType: 0
 --- !u!1001 &4087041631417843453
 PrefabInstance:
   m_ObjectHideFlags: 0
