+++ Assets/Prefabs/Environment/PlayerUnlockable/ShieldUnlock.prefab
@@ -10,7 +10,7 @@ GameObject:
   m_Component:
   - component: {fileID: 5320785212622099613}
   - component: {fileID: 6363545673026706925}
-  - component: {fileID: 6488219920590350764}
+  - component: {fileID: -8800628015503188254}
   m_Layer: 0
   m_Name: ShieldUnlock
   m_TagString: UnlockableShield
@@ -55,7 +55,7 @@ SphereCollider:
   serializedVersion: 3
   m_Radius: 0.75
   m_Center: {x: 0, y: 0, z: 0}
---- !u!114 &6488219920590350764
+--- !u!114 &-8800628015503188254
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -64,10 +64,10 @@ MonoBehaviour:
   m_GameObject: {fileID: 5523386548212966826}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: d05a0759fd7353d4ba4c2c3651978382, type: 3}
+  m_Script: {fileID: 11500000, guid: 747f8daff34ce5249af7ce003539deb8, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  speed: 1
+  abilityType: 1
 --- !u!1001 &49747167007877913
 PrefabInstance:
   m_ObjectHideFlags: 0
@@ -134,8 +134,31 @@ PrefabInstance:
     m_RemovedComponents: []
     m_RemovedGameObjects: []
     m_AddedGameObjects: []
-    m_AddedComponents: []
+    m_AddedComponents:
+    - targetCorrespondingSourceObject: {fileID: 3286663208650143064, guid: fe305a1575fbb7e47b2412543c318661,
+        type: 3}
+      insertIndex: -1
+      addedObject: {fileID: 4519200971851460178}
   m_SourcePrefab: {fileID: 100100000, guid: fe305a1575fbb7e47b2412543c318661, type: 3}
+--- !u!1 &3255027271533715009 stripped
+GameObject:
+  m_CorrespondingSourceObject: {fileID: 3286663208650143064, guid: fe305a1575fbb7e47b2412543c318661,
+    type: 3}
+  m_PrefabInstance: {fileID: 49747167007877913}
+  m_PrefabAsset: {fileID: 0}
+--- !u!114 &4519200971851460178
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 3255027271533715009}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: d05a0759fd7353d4ba4c2c3651978382, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  speed: 1
 --- !u!4 &8561264753710898370 stripped
 Transform:
   m_CorrespondingSourceObject: {fileID: 8538547988146206683, guid: fe305a1575fbb7e47b2412543c318661,
