+++ Assets/Programming/Character/OLD.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 95355594ce7cbc0498308d7625e59f1e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
