+++ Assets/Programming/Character/OLD/DamageIndicator.cs
@@ -1,7 +1,9 @@
+using System;
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
+[Obsolete("Please use the " + nameof(HurtIndicatorAuto) + " instead!", false)]
 public class DamageIndicator : MonoBehaviour
 {
     public float blinkIntensity;
