+++ Assets/Programming/Character/Player Combat and Abilities.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8c007fa0268d1344b8b2962772964289
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
