+++ Assets/Programming/Items.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3bd66a6d826987541967ba1593816559
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
