+++ Assets/Programming/Items/HealthObject.cs
@@ -1,0 +1,33 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class HealthObject : MonoBehaviour
+{
+    public float healAmount = 30f;
+
+
+    private void OnTriggerEnter(Collider other)
+    {
+        if (!other.CompareTag("Player")) return;
+
+        PlayerStats pStats = other.GetComponent<PlayerStats>();
+
+        if (pStats == null) return;
+
+        if (pStats.isDead) return;
+
+        if (pStats.PlayerHealth < pStats.PlayerMaxHealth)
+        {
+            Debug.Log("Healed player");
+
+            AudioManager.Instance.PlayAudio(false, false, pStats.audioSource, "Plr_Heal");
+
+            pStats.PlayerHealth += healAmount;
+
+            Destroy(gameObject);
+        }
+
+
+    }
+}
