+++ Assets/Programming/Items/PlayerAbilityUnlock.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: c20dfbcec6b638b418ef488a408cec0a
+guid: 747f8daff34ce5249af7ce003539deb8
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
